pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "django-rest-framework-missedcall"
version = "1.0.0"
description = "Production-grade flash-call authentication for Django REST Framework."
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
authors = [
    { name = "tabaro", email = "contact@tabaro.io" }
]
keywords = [
    "django", 
    "rest-framework", 
    "authentication", 
    "missed-call", 
    "flash-call", 
    "twilio"
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 3.2",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Security",
    "Topic :: Communications :: Telephony",
]
dependencies = [
    "django>=3.2",
    "djangorestframework>=3.12",
    "twilio>=7.0.0",
]

[project.urls]
"Homepage" = "https://github.com/tabaro/django-rest-framework-missedcall"
"Bug Tracker" = "https://github.com/tabaro/django-rest-framework-missedcall/issues"
"Documentation" = "https://github.com/tabaro/django-rest-framework-missedcall#readme"

[tool.setuptools]
packages = ["rest_framework_missedcall"]
include-package-data = true

[tool.setuptools.package-data]
"*" = ["*.po", "*.mo"]